Move the warm-probe gating test into the chat startup-gating suite
