+++ tests/test_chat_startup_gating.py
@@ -256,3 +256,18 @@ async def test_await_engine_builds_services_when_none_exist(_warming_services):
         ):
             assert await asyncio.to_thread(screen._await_chat_engine, widget) is True
         build.assert_called_once()
+
+
+def test_active_chat_warm_progress_skips_the_probe_when_nothing_is_warming():
+    """The task bar polls this every tick, so an idle TUI must not fire an HTTP
+    readiness probe forever -- the free in-process snapshot gates it."""
+    from lilbee.app.placement import active_chat_warm_progress
+
+    services = mock.MagicMock()
+    services.provider.warm_progress.return_value = None  # nothing warming
+    set_services(services)
+    try:
+        assert active_chat_warm_progress() is None
+        services.provider.role_ready.assert_not_called()
+    finally:
+        set_services(None)
